+++ tests/integration/models/reddit/test_subreddit.py
@@ -272,13 +272,6 @@ class TestSubredditFlair(IntegrationTest):
             mapping = self.subreddit.flair(redditor=self.REDDITOR)
             assert len(list(mapping)) == 1
 
-    def test__iter(self):
-        self.reddit.read_only = False
-        with self.recorder.use_cassette('TestSubredditFlair.test__iter'):
-            mapping = list(self.subreddit.flair)
-            assert len(mapping) > 0
-            assert all(isinstance(x['user'], Redditor) for x in mapping)
-
     def test_configure(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
@@ -867,13 +860,6 @@ class TestSubredditRelationships(IntegrationTest):
         with self.recorder.use_cassette('TestSubredditRelationships.banned'):
             self.add_remove(self.subreddit, self.REDDITOR, 'banned')
 
-    def test_banned__deprecated_iter(self):
-        self.reddit.read_only = False
-        banned = self.subreddit.banned
-        with self.recorder.use_cassette(
-                'TestSubredditRelationships.banned__deprecated_iter'):
-            assert len(list(banned)) > 0
-
     def test_banned__user_filter(self):
         self.reddit.read_only = False
         banned = self.subreddit.banned(redditor='pyapitestuser3')
@@ -886,13 +872,6 @@ class TestSubredditRelationships(IntegrationTest):
         with self.recorder.use_cassette(
                 'TestSubredditRelationships.contributor'):
             self.add_remove(self.subreddit, self.REDDITOR, 'contributor')
-
-    def test_contributor__deprecated_iter(self):
-        self.reddit.read_only = False
-        contributor = self.subreddit.contributor
-        with self.recorder.use_cassette(
-                'TestSubredditRelationships.contributor__deprecated_iter'):
-            assert len(list(contributor)) > 0
 
     @mock.patch('time.sleep', return_value=None)
     def test_contributor_leave(self, _):
@@ -918,13 +897,6 @@ class TestSubredditRelationships(IntegrationTest):
             # invite list.
             self.subreddit.moderator.add(self.REDDITOR)
             assert self.REDDITOR not in self.subreddit.moderator()
-
-    def test_moderator__deprecated_iter(self):
-        self.reddit.read_only = False
-        moderator = self.subreddit.moderator
-        with self.recorder.use_cassette(
-                'TestSubredditRelationships.moderator__deprecated_iter'):
-            assert len(list(moderator)) > 0
 
     @mock.patch('time.sleep', return_value=None)
     def test_moderator__limited_permissions(self, _):
@@ -992,13 +964,6 @@ class TestSubredditRelationships(IntegrationTest):
         with self.recorder.use_cassette('TestSubredditRelationships.muted'):
             self.add_remove(self.subreddit, self.REDDITOR, 'muted')
 
-    def test_muted__deprecated_iter(self):
-        self.reddit.read_only = False
-        muted = self.subreddit.muted
-        with self.recorder.use_cassette(
-                'TestSubredditRelationships.muted__deprecated_iter'):
-            assert len(list(muted)) > 0
-
     def test_moderator_remove_invite(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette('TestSubredditRelationships.'
@@ -1011,25 +976,11 @@ class TestSubredditRelationships(IntegrationTest):
                 'TestSubredditRelationships.wiki_banned'):
             self.add_remove(self.subreddit.wiki, self.REDDITOR, 'banned')
 
-    def test_wiki_banned__deprecated_iter(self):
-        self.reddit.read_only = False
-        wiki_banned = self.subreddit.wiki.banned
-        with self.recorder.use_cassette(
-                'TestSubredditRelationships.wiki_banned__deprecated_iter'):
-            assert len(list(wiki_banned)) > 0
-
     def test_wiki_contributor(self):
         self.reddit.read_only = False
         with self.recorder.use_cassette(
                 'TestSubredditRelationships.wiki_contributor'):
             self.add_remove(self.subreddit.wiki, self.REDDITOR, 'contributor')
-
-    def test_wiki_contributor__deprecated_iter(self):
-        self.reddit.read_only = False
-        wiki_contributor = self.subreddit.wiki.contributor
-        with self.recorder.use_cassette('TestSubredditRelationships.'
-                                        'wiki_contributor__deprecated_iter'):
-            assert len(list(wiki_contributor)) > 0
 
 
 class TestSubredditStreams(IntegrationTest):